Complete exercise 02 in intro-to-css
In this exercise, I made use of class and ID's
within HTML. Then in my style.css, I applied
styles to these classes and ID's. This exercise
makes you think hard on how to efficiently use
classes and ID's in order to keep code to a
minimum. Generally, the less code the better. This
improves readability for my future self and other
developers.